+++ packages/core/src/infrastructure/ai.ts
