+++ packages/core/src/modules/resource/use-cases/add-resource.ts
@@ -2,7 +2,7 @@ import { db } from '@oyster/db';
 import { id } from '@oyster/utils';
 
 import { job } from '@/infrastructure/bull';
-import { putObject } from '@/modules/object-storage';
+import { putObject } from '@/infrastructure/s3';
 import { type AddResourceInput } from '@/modules/resource/resource.types';
 
 export async function addResource(input: AddResourceInput) {
