+++ packages/core/src/modules/resume-review/resume-review.ts
@@ -6,13 +6,13 @@ import { z } from 'zod';
 import { db, relativeTime } from '@oyster/db';
 import { id } from '@oyster/utils';
 
+import { getChatCompletion } from '@/infrastructure/ai';
 import { job, registerWorker } from '@/infrastructure/bull';
 import {
   type GetBullJobData,
   ResumeReviewBullJob,
 } from '@/infrastructure/bull.types';
-import { getChatCompletion } from '@/modules/ai/ai';
-import { track } from '@/modules/mixpanel';
+import { track } from '@/infrastructure/mixpanel';
 import { ENV } from '@/shared/env';
 import { ColorStackError } from '@/shared/errors';
 import { fail, type Result, success } from '@/shared/utils/core.utils';
