+++ packages/core/src/modules/slack/slack.ts
@@ -6,15 +6,15 @@ import { z } from 'zod';
 
 import { type DB, db } from '@oyster/db';
 
-import { job } from '@/infrastructure/bull';
-import { getPineconeIndex } from '@/infrastructure/pinecone';
-import { cache, ONE_HOUR_IN_SECONDS } from '@/infrastructure/redis';
 import {
   createEmbedding,
   getChatCompletion,
   rerankDocuments,
-} from '@/modules/ai/ai';
-import { track } from '@/modules/mixpanel';
+} from '@/infrastructure/ai';
+import { job } from '@/infrastructure/bull';
+import { track } from '@/infrastructure/mixpanel';
+import { getPineconeIndex } from '@/infrastructure/pinecone';
+import { cache, ONE_HOUR_IN_SECONDS } from '@/infrastructure/redis';
 import { sendSlackNotification } from '@/modules/notification/use-cases/send-slack-notification';
 import { slack } from '@/modules/slack/instances';
 import { IS_PRODUCTION } from '@/shared/env';
